ArticuloUITEst
Se eliminan los archivos de prueba y se genera un solo Test
