+++ UnitTestPanaderia/UITestProject.csproj
@@ -58,10 +58,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ArticuloEliminarUITest.cs" />
-    <Compile Include="ArticuloEditarUITest.cs" />
-    <Compile Include="ArticuloCrearUITest.cs" />
-    <Compile Include="LoginUITest.cs" />
+    <Compile Include="ArticuloUITest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RecetaCrearUITest.cs" />
   </ItemGroup>
